:art: Inline a parent helper in Sync command
